+++ jschon/catalog/__init__.py
@@ -86,6 +86,10 @@ class Catalog:
         self._schema_cache: Dict[Hashable, Dict[URI, JSONSchema]] = {}
         self._enabled_formats: Set[str] = set()
 
+    def __repr__(self) -> str:
+        """Return `repr(self)`."""
+        return f'{self.__class__.__name__}({self.name!r})'
+
     def add_uri_source(self, base_uri: URI, source: Source):
         """Register a source for URI-identified JSON resources.
 
@@ -137,7 +141,7 @@ class Catalog:
 
         raise CatalogError(f'A source is not available for "{uri}"')
 
-    def create_vocabulary(self, uri: URI, *kwclasses: KeywordClass) -> None:
+    def create_vocabulary(self, uri: URI, *kwclasses: KeywordClass) -> Vocabulary:
         """Create a :class:`~jschon.vocabulary.Vocabulary` object, which
         may be used by a :class:`~jschon.vocabulary.Metaschema` to provide
         keyword classes used in schema construction.
@@ -145,8 +149,11 @@ class Catalog:
         :param uri: the URI identifying the vocabulary
         :param kwclasses: the :class:`~jschon.vocabulary.Keyword` classes
             constituting the vocabulary
+
+        :returns: the newly created :class:`~jschon.vocabulary.Vocabulary` instance
         """
         self._vocabularies[uri] = Vocabulary(uri, *kwclasses)
+        return self._vocabularies[uri]
 
     def get_vocabulary(self, uri: URI) -> Vocabulary:
         """Get a :class:`~jschon.vocabulary.Vocabulary` by its `uri`.
@@ -162,23 +169,33 @@ class Catalog:
     def create_metaschema(
             self,
             uri: URI,
-            core_vocabulary_uri: URI,
+            default_core_vocabulary_uri: URI = None,
             *default_vocabulary_uris: URI,
             **kwargs: Any,
-    ) -> None:
+    ) -> Metaschema:
         """Create, cache and validate a :class:`~jschon.vocabulary.Metaschema`.
 
         :param uri: the URI identifying the metaschema
-        :param core_vocabulary_uri: the URI identifying the metaschema's
-            core :class:`~jschon.vocabulary.Vocabulary`
+        :param default_core_vocabulary_uri: the URI identifying the metaschema's
+            core :class:`~jschon.vocabulary.Vocabulary`, used in the absence
+            of a ``"$vocabulary"`` keyword in the metaschema JSON file, or
+            if a known core vocabulary is not present under ``"$vocabulary"``
         :param default_vocabulary_uris: default :class:`~jschon.vocabulary.Vocabulary`
             URIs, used in the absence of a ``"$vocabulary"`` keyword in the
             metaschema JSON file
         :param kwargs: additional keyword arguments to pass through to the
             :class:`~jschon.jsonschema.JSONSchema` constructor
+
+        :returns: the newly created :class:`~jschon.vocabulary.Metaschema` instance
+
+        :raise CatalogError: if the metaschema is not valid
         """
         metaschema_doc = self.load_json(uri)
-        core_vocabulary = self.get_vocabulary(core_vocabulary_uri)
+        default_core_vocabulary = (
+            self.get_vocabulary(default_core_vocabulary_uri)
+            if default_core_vocabulary_uri
+            else None
+        )
         default_vocabularies = [
             self.get_vocabulary(vocab_uri)
             for vocab_uri in default_vocabulary_uris
@@ -186,13 +203,44 @@ class Catalog:
         metaschema = Metaschema(
             self,
             metaschema_doc,
-            core_vocabulary,
+            default_core_vocabulary,
             *default_vocabularies,
             **kwargs,
             uri=uri,
         )
         if not metaschema.validate().valid:
-            raise CatalogError("The metaschema is invalid against itself")
+            raise CatalogError(
+                "The metaschema is invalid against its own metaschema "
+                f'"{metaschema_doc["$schema"]}"'
+            )
+        return metaschema
+
+    def get_metaschema(self, uri: URI) -> Metaschema:
+        """Get a metaschema identified by `uri` from the ``'__meta__'`` cache, or
+        load it from configured sources if not already cached.
+
+        Note that metaschemas that do not declare a known core vocabulary
+        in ``"$vocabulary"`` must first be created using :meth:`create_metaschema`.
+
+        :param uri: the URI identifying the metaschema
+
+        :raise CatalogError: if the object referenced by `uri` is not
+            a :class:`~jschon.vocabulary.Metaschema`, or if it is not valid
+        :raise JSONSchemaError: if the metaschema is loaded from sources
+            but no known core vocabulary is present in ``"$vocabulary"``
+        """
+        try:
+            metaschema = self._schema_cache['__meta__'][uri]
+        except KeyError:
+            metaschema = None
+
+        if not metaschema:
+            metaschema = self.create_metaschema(uri)
+
+        if not isinstance(metaschema, Metaschema):
+            raise CatalogError(f"The schema referenced by {uri} is not a metaschema")
+
+        return metaschema
 
     def enable_formats(self, *format_attr: str) -> None:
         """Enable validation of the specified format attributes.