+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ Features:
 * JSON ``null``, ``true``, ``false`` literals
 * Relative JSON Pointer ``+``/``-`` array index adjustments
 * Unknown keywords are collected as annotations
+* Automatically create metaschemas as referenced by ``"$schema"``
+  on first access of the now-cached ``JSONSchema.metaschema`` property
+* Automatically detect the core vocabulary in metaschemas,
+  but allow specifying a default to use when none is detectable
 
 Experimental:
 
@@ -43,6 +47,11 @@ Miscellaneous:
 * Add ``JSONCompatible`` and ``Result`` classes to the top-level package API
 * Remove implicit fall-through to looking up a schema in the `__meta__` cache
   if not found in the parameterized cache, in ``Catalog.get_schema()`` (#40)
+* Added ``Catalog.get_metaschema()``, analogous to ``Catalog.get_schema()``
+* ``Catalog.create_metaschema()`` and ``Catalog.create_vocabulary()`` return the created instance
+* Rename ``core_vocabulary`` and ``core_vocabulary_uri`` parameters for
+  ``Metaschema.__init__()`` and ``Catalog.create_metaschema()`` respectively to
+  ``default_core_vocabulary`` and ``default_core_vocabulary_uri``
 * Improve kwarg-based construction of ``RelativeJSONPointer``
 * Allow passthrough of arguments to pytest when invoking tox
 * Add pytest command line options ``--testsuite-file`` and ``--testsuite-description``